+++ trunk/application/controllers/posts.php
@@ -47,10 +47,12 @@ class Posts extends MY_Controller
     
     public function add()
     {
+        if( !$this->userdata->is_admin() )
+            redirect(base_url());
         parent::add('add_post', $this->router->class, $this->router->method);
         
         $data = array(
-            'error'         => $this->form_validation->form_required(array('title','content'))
+            'error'         => $this->form_validation->form_required(array('title','content','priority'))
         );
         
         $this->load->view('container', array_merge($this->data, $data));
@@ -58,6 +60,10 @@ class Posts extends MY_Controller
     
     public function edit( $post_id )
     {
+        if( $post_id == '')
+            redirect('404');
+        if( !$this->userdata->is_admin() )
+            redirect(base_url());
         parent::edit('edit_posts', $post_id, $this->router->class, $this->router->method);
         
         $this->load->model('selecter');
@@ -65,7 +71,7 @@ class Posts extends MY_Controller
         
         $data = array(
             'post_id'       => $post_id,
-            'error'         => $this->form_validation->form_required(array('title','content'))
+            'error'         => $this->form_validation->form_required(array('title','content','priority'))
         );
         
         $this->load->view('container', array_merge($this->data, $data));
@@ -73,13 +79,17 @@ class Posts extends MY_Controller
     
     public function delete( $post_id )
     {
+        if( $post_id == '')
+            redirect('404');
+        if( !$this->userdata->is_admin() )
+            redirect(base_url());
         parent::delete('remove_post', $post_id, $this->router->class);
             
         $data = array(
             'view'            => 'confirm_view',
             'type'            => 'delete',
             'langs'           => array($this->lang->line('confirm_yes'), $this->lang->line('confirm_no')),
-            'method'          => $this->router->class.'/'.$post_id
+            'method'          => $this->router->class.'/'.$this->router->method.'/'.$post_id
         );
             
         $this->load->view('container', array_merge($this->data, $data));
